+++ uikit/src/main/java/com/sendbird/uikit/internal/extensions/MessageListComponentExtensions.kt
@@ -1,0 +1,10 @@
+package com.sendbird.uikit.internal.extensions
+import com.sendbird.uikit.internal.interfaces.OnSubmitButtonClickListener
+import com.sendbird.uikit.modules.components.MessageListComponent
+
+private var submitButtonClickListener: OnSubmitButtonClickListener? = null
+internal var MessageListComponent.submitButtonClickListener: OnSubmitButtonClickListener?
+    get() = com.sendbird.uikit.internal.extensions.submitButtonClickListener
+    set(value) {
+        com.sendbird.uikit.internal.extensions.submitButtonClickListener = value
+    }