+++ uikit/src/main/java/com/sendbird/uikit/internal/interfaces/OnSubmitButtonClickListener.kt
@@ -1,0 +1,8 @@
+package com.sendbird.uikit.internal.interfaces
+
+import com.sendbird.android.message.BaseMessage
+import com.sendbird.uikit.internal.model.Form
+
+internal fun interface OnSubmitButtonClickListener {
+    fun onClicked(message: BaseMessage, form: Form)
+}