+++ uikit/src/main/java/com/sendbird/uikit/internal/ui/viewholders/FormMessageViewHolder.kt
@@ -1,0 +1,29 @@
+package com.sendbird.uikit.internal.ui.viewholders
+
+import android.view.View
+import com.sendbird.android.channel.BaseChannel
+import com.sendbird.android.message.BaseMessage
+import com.sendbird.android.message.UserMessage
+import com.sendbird.uikit.activities.viewholder.MessageViewHolder
+import com.sendbird.uikit.databinding.SbViewFormMessageBinding
+import com.sendbird.uikit.internal.extensions.forms
+import com.sendbird.uikit.internal.interfaces.OnSubmitButtonClickListener
+import com.sendbird.uikit.model.MessageListUIParams
+
+internal class FormMessageViewHolder internal constructor(
+    val binding: SbViewFormMessageBinding,
+    messageListUIParams: MessageListUIParams
+) : MessageViewHolder(binding.root, messageListUIParams) {
+    var onSubmitClickListener: OnSubmitButtonClickListener? = null
+    override fun bind(channel: BaseChannel, message: BaseMessage, messageListUIParams: MessageListUIParams) {
+        if (message !is UserMessage) return
+        val form = message.forms.firstOrNull() ?: return
+        binding.formsMessageView.messageUIConfig = messageUIConfig
+        binding.formsMessageView.drawFormMessage(message, messageListUIParams)
+        binding.formsMessageView.setSubmitButtonClickListener {
+            onSubmitClickListener?.onClicked(message, form)
+        }
+    }
+
+    override fun getClickableViewMap(): Map<String, View> = mapOf()
+}